+++ apps/docs/demos/radio-group/usage/App.tsx
@@ -1,11 +1,28 @@
-import { RadioGroup, RadioGroupItem } from "@optiaxiom/react";
+import type { ComponentPropsWithoutRef } from "react";
 
-export function App() {
+import { Field, RadioGroup, RadioGroupItem } from "@optiaxiom/react";
+
+export function App({
+  description,
+  error,
+  label = "Input label",
+  required = false,
+}: Pick<
+  ComponentPropsWithoutRef<typeof Field>,
+  "description" | "error" | "label" | "required"
+>) {
   return (
-    <RadioGroup>
-      <RadioGroupItem value="one">Option One</RadioGroupItem>
-      <RadioGroupItem value="two">Option Two</RadioGroupItem>
-      <RadioGroupItem value="three">Option Three</RadioGroupItem>
-    </RadioGroup>
+    <Field
+      description={description}
+      error={error}
+      label={label}
+      required={required}
+    >
+      <RadioGroup>
+        <RadioGroupItem value="one">Option One</RadioGroupItem>
+        <RadioGroupItem value="two">Option Two</RadioGroupItem>
+        <RadioGroupItem value="three">Option Three</RadioGroupItem>
+      </RadioGroup>
+    </Field>
   );
 }
