+++ apps/docs/demos/textarea/usage/App.tsx
@@ -1,5 +1,24 @@
-import { Textarea } from "@optiaxiom/react";
+import type { ComponentPropsWithoutRef } from "react";
 
-export function App() {
-  return <Textarea placeholder="Enter text..." />;
+import { Field, Textarea } from "@optiaxiom/react";
+
+export function App({
+  description,
+  error,
+  label = "Input label",
+  required = false,
+}: Pick<
+  ComponentPropsWithoutRef<typeof Field>,
+  "description" | "error" | "label" | "required"
+>) {
+  return (
+    <Field
+      description={description}
+      error={error}
+      label={label}
+      required={required}
+    >
+      <Textarea placeholder="Enter text..." />
+    </Field>
+  );
 }
