+++ apps/docs/pages/components/input.mdx
@@ -32,12 +32,6 @@ Enable the `readOnly` prop to display the read-only state of the input field.
 
 <ax-demo name="input/readonly-usage" />
 
-### Accessibility
-
-Combine with the `Field` component to provide context such as label/help/required etc.
-
-<ax-demo name="input/label-usage" />
-
 ## Addons
 
 ### Left and right addons
