+++ apps/docs/pages/components/textarea.mdx
@@ -26,12 +26,6 @@ Enable the `readOnly` prop to display the read-only state of the text field.
 
 <ax-demo name="textarea/readonly-usage" />
 
-### Accessibility
-
-Combine with the `Field` component to provide context such as label/help/required etc.
-
-<ax-demo name="textarea/label-usage" />
-
 ## Sizing
 
 ### Fixed size
